modified the tables

+++ sqlTables/employeeTable.sql
@@ -4,7 +4,7 @@ use bar;
 create table if not exists employee(
 e_id int(5) not null,
 ename varchar(45) not null,
-salary int(10),
+salary float,
 datehired varchar(10),
 primary key(e_id));
